Wait until pegin providers are loaded
